+++ src/contracts/index.ts
@@ -56,3 +56,4 @@ export * from "./EventContractObservation";
 export * from "./EventContractShadowLedger";
 export * from "./ResearchIntegrity";
 export * from "./ResearchDatasetQualification";
+export * from "./ResearchShadowDatasetAssembly";